requirements.txt
streamlit==1.44.0
streamlit_folium==0.24.1
streamlit_option_menu==0.4.0
mysql-connector-python==9.2.0
st-gsheets-connection==0.1.0
vega_datasets==0.9.0
prettytable==3.16.0
matplotlib==3.10.1
geopandas==1.0.1
Authlib==1.5.2
folium==0.19.5
pandas==2.2.3
numpy==2.2.4
langchain==0.3.21
faiss-cpu==1.11.0
sentence-transformers==4.1.0
unstructured==0.17.2
python-docx==1.1.2
openai==1.76.0